FTP Sender doesn't send empty files, closes #722

+++ app/models/ftp_sender.rb
@@ -11,29 +11,36 @@ class FtpSender
   def self.send_file server, username, password, file, opts = {}
     log = ["Attempting to send FTP file to #{server} with username #{username}."]
     begin
-      my_opts = {:binary=>true,:passive=>true,:remote_file_name=>File.basename(file)}.merge opts
+      my_opts = {:binary=>true,:passive=>true,:remote_file_name=>File.basename(file),
+        :force_empty=>false}.merge opts
       remote_name = my_opts[:remote_file_name]
-      Net::FTP.open(server) do |f|
-        log << "Opened connection to #{server}"
-        f.login(username,password)
-        log << "Logged in with account #{username}"
-        f.passive = my_opts[:passive]
-        log << "Set passive to #{my_opts[:passive]}"
-        if my_opts[:folder]
-          f.chdir(my_opts[:folder]) 
-          log << "Changed to folder #{my_opts[:folder]}" 
+      write_file_to_db = false
+      if File.new(file).size > 0 || my_opts[:force_empty]
+        write_file_to_db = true
+        Net::FTP.open(server) do |f|
+          log << "Opened connection to #{server}"
+          f.login(username,password)
+          log << "Logged in with account #{username}"
+          f.passive = my_opts[:passive]
+          log << "Set passive to #{my_opts[:passive]}"
+          if my_opts[:folder]
+            f.chdir(my_opts[:folder]) 
+            log << "Changed to folder #{my_opts[:folder]}" 
+          end
+          data = nil
+          if my_opts[:binary]
+            data = File.open(file, "rb") {|f| f.read}
+            f.putbinaryfile file, remote_name
+            log << "Put binary file #{file.to_s} as #{remote_name}"
+          else
+            f.puttextfile file, remote_name
+            log << "Put text file #{file.to_s} as #{remote_name}"
+            data = File.open(file, "r") {|f| f.read}
+          end
+          log << "Session completed successfully."
         end
-        data = nil
-        if my_opts[:binary]
-          data = File.open(file, "rb") {|f| f.read}
-          f.putbinaryfile file, remote_name
-          log << "Put binary file #{file.to_s} as #{remote_name}"
-        else
-          f.puttextfile file, remote_name
-          log << "Put text file #{file.to_s} as #{remote_name}"
-          data = File.open(file, "r") {|f| f.read}
-        end
-        log << "Session completed successfully."
+      else
+        log << "File was empty, not sending."
       end
     rescue
       log << "ERROR: #{$!.message}"
@@ -43,7 +50,7 @@ class FtpSender
                         :server => server,
                         :file_name => File.basename(file),
                         :log => log.join("\n"),
-                        :data => File.open(file, "rb") { |f| f.read })
+                        :data => (write_file_to_db ? File.open(file, "rb") { |f| f.read } : nil))
     end
   end
